Added a flag for accessing the LOTR Mod
PvpMode now will only access the hired unit data of the LOTR Mod if it didn't fail before. Every exception catched here indicates a serious issue which was probably caused by changes in the LOTR Mod which means that they'll occurr every time we try to access the hired unit data - if if failed once, it'll fail again and again. This improves the performance of PvpMode and prevents console spam.